third commit

+++ strategy/strategy.go
@@ -2,13 +2,13 @@ package strategy
 
 import "fmt"
 
-// 销售策略
+// SellStrategy 销售策略
 type SellStrategy interface {
-	//根据原价得到售卖价
+	// GetPrice 根据原价得到售卖价
 	GetPrice(price float64) float64
 }
 
-// 销售策略A，按八折出售
+// StrategyA 销售策略A，按八折出售
 type StrategyA struct {
 }
 
@@ -16,7 +16,7 @@ func (s *StrategyA) GetPrice(price float64) float64 {
 	return price * 0.8
 }
 
-// 销售策略B，满200-100
+// StrategyB 销售策略B，满200-100
 type StrategyB struct {
 }
 
@@ -27,7 +27,7 @@ func (s *StrategyB) GetPrice(price float64) float64 {
 	return price
 }
 
-// 环境
+// Goods 环境
 type Goods struct {
 	Price    float64
 	strategy SellStrategy
@@ -39,5 +39,5 @@ func (goods *Goods) SetStrategy(s SellStrategy) {
 
 func (goods *Goods) SellGoods() float64 {
 	fmt.Println("物品原价为:", goods.Price, "元")
-	return goods.strategy.GetPrice(float64(goods.Price))
+	return goods.strategy.GetPrice(goods.Price)
 }
